WIP: Track processes using pidfds

+++ lmkd/lmkd.c
@@ -121,7 +121,17 @@
 
 #define FAIL_REPORT_RLIMIT_MS 1000
 
+#define __NR_pidfd_send_signal 424
 #define __NR_pidfd_open 434
+
+static inline int sys_pidfd_open(pid_t pid, unsigned int flags) {
+    return syscall(__NR_pidfd_open, pid, flags);
+}
+
+static inline int sys_pidfd_send_signal(int pidfd, int sig, siginfo_t *info,
+                                        unsigned int flags) {
+    return syscall(__NR_pidfd_send_signal, pidfd, sig, info, flags);
+}
 
 /* default to old in-kernel interface if no memory pressure events */
 static bool use_inkernel_interface = true;
@@ -186,7 +196,7 @@ enum polling_update {
     POLLING_START,
     POLLING_STOP,
     POLLING_PAUSE,
-    POLLING_CONTINUE,
+    POLLING_RESUME,
 };
 
 struct polling_params {
@@ -454,6 +464,7 @@ struct adjslot_list {
 struct proc {
     struct adjslot_list asl;
     int pid;
+    int pidfd;
     uid_t uid;
     int oomadj;
     struct proc *pidhash_next;
@@ -657,7 +668,7 @@ static void proc_insert(struct proc *procp) {
     proc_slot(procp);
 }
 
-static int pid_remove(int pid) {
+static int pid_remove(int pid, bool close_pidfd) {
     int hval = pid_hashfn(pid);
     struct proc *procp;
     struct proc *prevp;
@@ -675,6 +686,9 @@ static int pid_remove(int pid) {
         prevp->pidhash_next = procp->pidhash_next;
 
     proc_unslot(procp);
+    if (close_pidfd && procp->pidfd >= 0) {
+        close(procp->pidfd);
+    }
     free(procp);
     return 0;
 }
@@ -844,6 +858,16 @@ static void cmd_procprio(LMKD_CTRL_PACKET packet) {
 
     procp = pid_lookup(params.pid);
     if (!procp) {
+            int pidfd = -1;
+
+            if (pidfd_supported) {
+                pidfd = TEMP_FAILURE_RETRY(sys_pidfd_open(params.pid, 0));
+                if (pidfd < 0) {
+                    ALOGE("Attempt to register a dead task (pid %d)", params.pid);
+                    return;
+                }
+            }
+
             procp = malloc(sizeof(struct proc));
             if (!procp) {
                 // Oh, the irony.  May need to rebuild our state.
@@ -851,6 +875,7 @@ static void cmd_procprio(LMKD_CTRL_PACKET packet) {
             }
 
             procp->pid = params.pid;
+            procp->pidfd = pidfd;
             procp->uid = params.uid;
             procp->oomadj = params.oomadj;
             proc_insert(procp);
@@ -873,7 +898,7 @@ static void cmd_procremove(LMKD_CTRL_PACKET packet) {
      * WARNING: After pid_remove() procp is freed and can't be used!
      * Therefore placed at the end of the function.
      */
-    pid_remove(params.pid);
+    pid_remove(params.pid, true);
 }
 
 static void cmd_procpurge() {
@@ -1658,7 +1683,7 @@ static struct proc *proc_get_heaviest(int oomadj) {
         int tasksize = proc_get_size(pid);
         if (tasksize <= 0) {
             struct adjslot_list *next = curr->next;
-            pid_remove(pid);
+            pid_remove(pid, true);
             curr = next;
         } else {
             if (tasksize > maxsize) {
@@ -1704,10 +1729,6 @@ static void set_process_group_and_prio(int pid, SchedPolicy sp, int prio) {
         }
     }
     closedir(d);
-}
-
-static inline int sys_pidfd_open(pid_t pid, unsigned int flags) {
-	return syscall(__NR_pidfd_open, pid, flags);
 }
 
 static bool is_kill_pending(void) {
@@ -1769,10 +1790,10 @@ static void stop_wait_for_proc_kill(bool finished) {
 
 static void kill_done_handler(int data __unused, uint32_t events __unused, struct polling_params *poll_params) {
     stop_wait_for_proc_kill(true);
-    poll_params->update = POLLING_CONTINUE;
+    poll_params->update = POLLING_RESUME;
 }
 
-static void start_wait_for_proc_kill(int pid) {
+static void start_wait_for_proc_kill(int pid_of_fd) {
     static struct event_handler_info kill_done_hinfo = { 0, kill_done_handler };
     struct epoll_event epev;
 
@@ -1782,15 +1803,10 @@ static void start_wait_for_proc_kill(int pid) {
         stop_wait_for_proc_kill(false);
     }
 
-    if (!pidfd_supported) {
-        /* If pidfd is not supported store PID of the process being killed */
-        last_kill_pid_or_fd = pid;
-        return;
-    }
+    last_kill_pid_or_fd = pid_of_fd;
 
-    last_kill_pid_or_fd = TEMP_FAILURE_RETRY(sys_pidfd_open(pid, 0));
-    if (last_kill_pid_or_fd < 0) {
-        ALOGE("pidfd_open for process pid %d failed; errno=%d", pid, errno);
+    if (!pidfd_supported) {
+        /* If pidfd is not supported just store PID and exit */
         return;
     }
 
@@ -1809,6 +1825,7 @@ static void start_wait_for_proc_kill(int pid) {
 static int kill_one_process(struct proc* procp, int min_oom_score, const char *reason,
                             struct timespec *tm) {
     int pid = procp->pid;
+    int pidfd = procp->pidfd;
     uid_t uid = procp->uid;
     int tgid;
     char *taskname;
@@ -1852,11 +1869,14 @@ static int kill_one_process(struct proc* procp, int min_oom_score, const char *r
 
     TRACE_KILL_START(pid);
 
-    /* Have to start waiting before sending SIGKILL to make sure pid is valid */
-    start_wait_for_proc_kill(pid);
-
     /* CAP_KILL required */
-    r = kill(pid, SIGKILL);
+    if (pidfd < 0) {
+        start_wait_for_proc_kill(pid);
+        r = kill(pid, SIGKILL);
+    } else {
+        start_wait_for_proc_kill(pidfd);
+        r = sys_pidfd_send_signal(pidfd, SIGKILL, NULL, 0);
+    }
 
     set_process_group_and_prio(pid, SP_FOREGROUND, ANDROID_PRIORITY_HIGHEST);
 
@@ -1897,7 +1917,7 @@ out:
      * WARNING: After pid_remove() procp is freed and can't be used!
      * Therefore placed at the end of the function.
      */
-    pid_remove(pid);
+    pid_remove(pid, false);
     return result;
 }
 
@@ -2108,14 +2128,7 @@ static void mp_event_psi(int data, uint32_t events, struct polling_params *poll_
 
     kill_pending = is_kill_pending();
     if (kill_pending) {
-        if (is_waiting_for_kill()) {
-            /* pause polling if we are waiting for process death notification */
-            poll_params->update = POLLING_PAUSE;
-            return;
-        }
-        /* kernel does not support pidfd waiting, start quick polling instead */
-        poll_params->polling_interval_ms = PSI_POLL_PERIOD_SHORT_MS;
-        goto no_kill;
+        goto killing;
     }
 
     if (vmstat_parse(&vs) < 0) {
@@ -2243,9 +2256,9 @@ do_kill:
     if (kill_reason != NONE) {
         find_and_kill_process(0, kill_desc, &curr_tm);
         last_kill_memfree = vs.field.nr_free_pages;
-        poll_params->polling_interval_ms = PSI_POLL_PERIOD_SHORT_MS;
         just_killed = true;
         meminfo_log(&mi);
+        goto killing;
     }
 
 no_kill:
@@ -2256,6 +2269,17 @@ no_kill:
      * without causing memory pressure
      */
     if (events || just_killed || reclaim == DIRECT_RECLAIM) {
+        poll_params->update = POLLING_START;
+    }
+    return;
+
+killing:
+    if (is_waiting_for_kill()) {
+        /* pause polling if we are waiting for process death notification */
+        poll_params->update = POLLING_PAUSE;
+    } else {
+        /* kernel does not support pidfd waiting, start quick polling instead */
+        poll_params->polling_interval_ms = PSI_POLL_PERIOD_SHORT_MS;
         poll_params->update = POLLING_START;
     }
 }
@@ -2817,18 +2841,31 @@ static void mainloop(void) {
         int i;
 
         if (poll_params.poll_handler) {
-            /* Calculate next timeout */
-            clock_gettime(CLOCK_MONOTONIC_COARSE, &curr_tm);
-            delay = get_time_diff_ms(&poll_params.last_poll_tm, &curr_tm);
-            delay = (delay < poll_params.polling_interval_ms) ?
-                poll_params.polling_interval_ms - delay : poll_params.polling_interval_ms;
+            bool poll_now;
 
-            /* Wait for events until the next polling timeout */
-            nevents = epoll_wait(epollfd, events, maxevents, delay);
+            if (poll_params.poll_handler == poll_params.paused_handler) {
+                /*
+                 * Just transitioned into POLLING_RESUME. Reset paused_handler
+                 * and poll immediately
+                 */
+                poll_params.paused_handler = NULL;
+                poll_now = true;
+                nevents = 0;
+            } else {
+                /* Calculate next timeout */
+                clock_gettime(CLOCK_MONOTONIC_COARSE, &curr_tm);
+                delay = get_time_diff_ms(&poll_params.last_poll_tm, &curr_tm);
+                delay = (delay < poll_params.polling_interval_ms) ?
+                    poll_params.polling_interval_ms - delay : poll_params.polling_interval_ms;
 
-            clock_gettime(CLOCK_MONOTONIC_COARSE, &curr_tm);
-            if (get_time_diff_ms(&poll_params.last_poll_tm, &curr_tm) >=
-                poll_params.polling_interval_ms) {
+                /* Wait for events until the next polling timeout */
+                nevents = epoll_wait(epollfd, events, maxevents, delay);
+
+                clock_gettime(CLOCK_MONOTONIC_COARSE, &curr_tm);
+                poll_now = (get_time_diff_ms(&poll_params.last_poll_tm, &curr_tm) >=
+                    poll_params.polling_interval_ms);
+            }
+            if (poll_now) {
                 /* Set input params for the call */
                 poll_params.poll_handler->handler(poll_params.poll_handler->data, 0, &poll_params);
                 poll_params.last_poll_tm = curr_tm;
@@ -2845,7 +2882,7 @@ static void mainloop(void) {
                         poll_params.paused_handler = poll_params.poll_handler;
                         poll_params.poll_handler = NULL;
                         break;
-                    case (POLLING_CONTINUE):
+                    case (POLLING_RESUME):
                         /* Already polling, ignore */
                         break;
                     default: /* POLLING_NO_CHANGE */
@@ -2920,9 +2957,8 @@ static void mainloop(void) {
                         poll_params.paused_handler = handler_info;
                         poll_params.poll_handler = NULL;
                         break;
-                    case (POLLING_CONTINUE):
+                    case (POLLING_RESUME):
                         poll_params.poll_handler = poll_params.paused_handler;
-                        poll_params.poll_handler->handler(poll_params.poll_handler->data, 0, &poll_params);
                         clock_gettime(CLOCK_MONOTONIC_COARSE, &curr_tm);
                         poll_params.poll_start_tm = poll_params.last_poll_tm = curr_tm;
                         break;
